Return type changes for WB-MCM8

+++ src/device/wb_mcm8.h
@@ -11,10 +11,10 @@ namespace EDWB
         MCM8(ModbusClient& client) : Device(client) {}
 
         bool getInputState(uint8_t pin);
-        uint16_t getShortPressCount(uint8_t pin);
-        uint16_t getLongPressCount(uint8_t pin);
-        uint16_t getDoublePressCount(uint8_t pin);
-        uint16_t getShortAndThenLongPressCount(uint8_t pin);
+        int32_t getShortPressCount(uint8_t pin);
+        int32_t getLongPressCount(uint8_t pin);
+        int32_t getDoublePressCount(uint8_t pin);
+        int32_t getShortAndThenLongPressCount(uint8_t pin);
 
         bool setInputMode(uint8_t pin, uint16_t mode);
     };
